Fix (cookie.jsx: 지저분한 자잘구리들 정리, othersPage.jsx & modifyProfime.jsx & follow.jsx: 각 컴포넌트 폴더 경로 이동에 따른 모듈 경로 변경), Delete(signUpForm.jsx), Rename(src/components/features/myPage/myPageFollow.jsx -> src/components/features/follow/MyPageFollow.jsx, src/components/features/myPage/UserProfileContainer.jsx -> src/components/features/othersPage/UserProfileContainer.jsx, src/components/features/myPage/ProfileModifyForm.jsx -> src/components/modifyProfile/ProfileModifyForm.jsx)
